cloud_orchestration/requirements.txt
twelve-data
python-telegram-bot
supabase
fastapi
uvicorn
python-dotenv
pytest
httpx
pydantic
google-generativeai
